requirements.txt
anyio==3.6.2
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.5
dnspython==2.2.1
email-validator==1.3.0
et-xmlfile==1.1.0
fastapi==0.85.1
FastAPI-SQLAlchemy==0.2.1
greenlet==1.1.3.post0
h11==0.14.0
httptools==0.5.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
openpyxl==3.0.10
orjson==3.8.0
psycopg2-binary==2.9.5
pydantic==1.10.2
python-dotenv==0.21.0
python-multipart==0.0.5
PyYAML==6.0
requests==2.28.1
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.42
starlette==0.20.4
typing_extensions==4.4.0
ujson==5.5.0
urllib3==1.26.12
uvicorn==0.18.3
watchfiles==0.18.0
websockets==10.3
